Partially completed the exception handling

+++ src/main/java/com/fyndus/schoolmanagement/service/CourseService.java
@@ -1,7 +1,9 @@
 package com.fyndus.schoolmanagement.service;
 
 import com.fyndus.schoolmanagement.entity.Course;
+import com.fyndus.schoolmanagement.DTO.ResponseDTO;
 import com.fyndus.schoolmanagement.repository.CourseRepository;
+import com.fyndus.schoolmanagement.util.ResponseMessage;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
@@ -16,34 +18,65 @@ public class CourseService {
         this.courseRepo = courseRepo;
     }
 
-    public Course createCourse(Course course) {
+    public ResponseDTO createCourse(Course course) {
         course.setCreatedAt(Instant.now());
-        return this.courseRepo.save(course);
+        return ResponseDTO.builder().data(this.courseRepo.save(course)).message(course.getName()+" "+ResponseMessage.CREATED).build();
     }
 
-    public Course findById(Long courseId) {
-        return this.courseRepo.findById(courseId).orElse(null);
+    public ResponseDTO findById(Long courseId) {
+        final Course course;
+        try {
+            course = this.courseRepo.findById(courseId).orElseThrow(NullPointerException::new);
+        }catch (NullPointerException e) {
+            return ResponseDTO.builder().message(ResponseMessage.NOT_FOUND).build();
+        }
+        return ResponseDTO.builder().data(course).message(ResponseMessage.FOUND).build();
     }
 
-    public List<Course> findAll() {
-        return this.courseRepo.findAll();
+    public ResponseDTO findAll() {
+        final List<Course> courses = this.courseRepo.findAll();
+        if(courses.isEmpty()) {
+            return ResponseDTO.builder().message(ResponseMessage.NOT_FOUND).build();
+        }
+        return ResponseDTO.builder()
+         .data(courses)
+         .message(ResponseMessage.FOUND)
+         .build();
+//        ResponseDTO responseDTO = new ResponseDTO();
+//        responseDTO.setData(courses);
+//        responseDTO.setMessage(ResponseMessage.FOUND);
     }
 
-    public Course updateCourse(Long courseId, Course course) {
-        final Course temp = this.courseRepo.findById(courseId).orElse(null);
-        if(temp == null) return temp;
+    public ResponseDTO updateCourse(Long courseId, Course course) {
+        final Course temp;
+        try {
+            temp = this.courseRepo.findById(courseId).orElseThrow(NullPointerException::new);
+        } catch(NullPointerException e) {
+            return ResponseDTO.builder().message(course.getName() + "-" + ResponseMessage.NOT_FOUND).build();
+        }
         temp.setUpdatedAt(Instant.now());
         temp.setName(course.getName());
-        return this.courseRepo.save(temp);
+        return ResponseDTO.builder().data(this.courseRepo.save(temp)).message(ResponseMessage.UPDATED).build();
     }
 
-    public String deleteAll() {
+    public ResponseDTO deleteAll() {
+        final List<Course> courses = this.courseRepo.findAll();
+        if(courses.isEmpty()) {
+            return ResponseDTO.builder().message(ResponseMessage.EMPTY).build();
+        }
+        final ResponseDTO responseDTO = ResponseDTO.builder().data(courses).message(ResponseMessage.DELETED).build();
         this.courseRepo.deleteAll();
-        return "All courses deleted";
+        return responseDTO;
     }
 
-    public String deleteById(Long courseId) {
+    public ResponseDTO deleteById(Long courseId) {
+        final ResponseDTO responseDTO;
+        try {
+            responseDTO = ResponseDTO.builder().data(this.courseRepo.findById(courseId).orElseThrow(NullPointerException::new)).message("Course with id: "+courseId+" "+ResponseMessage.DELETED).build();
+        } catch (NullPointerException e) {
+            return ResponseDTO.builder().message(ResponseMessage.NOT_FOUND).build();
+        }
         this.courseRepo.deleteById(courseId);
-        return "Course with id: "+courseId+" deleted";
+        return responseDTO;
     }
 }
